Part 1 was super easy.
This includes an upgrade to manhattanDistance allowing the user to
specify the dimensional length of the arrays being passed. This allows
the user to pass arrays longer than what should be evaluated for
distance purposes (for example, an array of the format [x, y, z,
strength] could be passed, and manhattanDistance could be instructed to
only evaluate the first three dimensions).