2018/2018
Day 1, 2018.

Day 2, 2018. (#1)

* Setup, input, part 1.

* Part 2.

Day 3, 2018. (#2)

* Input, setup, part 1.

* Part 2.

Day 4, 2018. (#3)

* Input.

* Part 1 - Sleepiest guard's sleepiest minute.

* Part 2 - Overall sleepiest guard-minute.

Day 5, 2018. (#4)

* Set up input. React polymers (part 1).

* Filter polymer repeatedly to find the best filter.

Day 6, 2018. (#5)

* Parse input and display map.

* Find largest non-infinite area (part 1).

* Find locations under a target total distance from all coords (part 2).

Day 7, 2018. (#6)

* Input. Task order, 1 worker, no time. (Part 1)

* Calculate total task time, with workers (part 2).

Day 8, 2018. (#7)

* Input. Calculate checksum. (Part 1)

* Calculate root node value. (Part 2)

Day 9, 2018. (#8)

Parts 1 and 2 in one go!

Day 10, 2018. (#9)

* Parse input.

* Render the display, if it's small enough.

* Find the message! (Parts 1 and 2)

Day 11, 2018. (#10)

* Calculate 3x3 cell grid power (part 1).

This one was pretty simple and there’s nothing fancy to display.

* Part 2 - variable grid sizes.

For this one, it takes forever to run normally, so you need to tell
it to stop checking bigger ones from the same grid square if each
one is getting lesser.

Day 12, 2018. (#11)

* Parse input.

* Advance generation (part 1).

* For part 2, detect a loop and break out of it.

Examine what's happening in the loop and you can predict the end.

Day 13, 2018. (#12)

* Parse and display input.

* Advance one tick.

* Detect the first collision (part 1).

* Remove on collision and find final cart (part 2).

Day 14, 2018. (#13)

* Initial setup and display.

* Create recipes and advance.

* Find good recipes (part 1).

* Find good recipes (part 2).

This runs kinda slowly but will get you the correct answer.

Day 15, 2018. (#14)

* Parse and display input, fight combat (part 1).

* Part 2 wasn't very difficult after Part 1!

Day 16, 2018. (#15)

* This part was pretty simple. Part 2 might be more challenging!

* Step 1 - How many ops does each sample match?

(More granular than just three or more)

* Step 2 - Actually figure out which ops are which

* Step 3 - Actually run the steps and get the answer!

Day 17, 2018. (#16)

* Read and render the input, flow the water.

* Part 2 was super easy.

Day 18, 2018. (#17)

* Parse and render the input data.

* This one was pretty simple (so far!!)

* Look for previously seen field states.

A given input will always produce the same output, so if we see a board
state that we've seen before, we know we're in a loop and we can react
accordingly by just skipping ahead to the appropriate time.

Day 19, 2018. (#18)

This will run the program...

... It is left as an exercise to the reader to actually figure out what
the result will be without running the full program in its entirety (it
will loop for a very long time if you do!).

Day 20, 2018. (#19)

* Allow input data selection.

* Solve part 1!

This wasn't my first idea/attempt but the first one didn't work out. :P

* Part 2 was super easy.

* Let's actually render the map.

Day 21, 2018. (#20)

* Find first exit code.

* Find all the exit codes. (Part 2)

I don't like that the only way to solve this is to just let the program
run for a really long time, but I think it really is the only way.

Day 22, 2018. (#21)

* Initial setup and input.

* Another pretty simple one.

* Display map.

* Solve part 2!

Day 23, 2018. (#22)

* Part 1 was super easy.

This includes an upgrade to manhattanDistance allowing the user to
specify the dimensional length of the arrays being passed. This allows
the user to pass arrays longer than what should be evaluated for
distance purposes (for example, an array of the format [x, y, z,
strength] could be passed, and manhattanDistance could be instructed to
only evaluate the first three dimensions).

* Part 2.

Day 24, 2018. (#23)

* Initial input and parsing.

* Run combat!!

* Part 2 - apply boost, detect stalemate.

Day 25, 2018. (#24)

* Setup. Read and display input.

* Good constellation count.