2022/day 07 (#83)
* Part 1 in progress.

Right answer on demo, wrong answer on real. I think the problem is that
I'm not accounting for directories at different nesting levels with the
same name, ie `/aa`, `/bb/aa`, `/cc/bb/aa`, etc. That's just a guess at
the bug though. If I start accounting for that and still get it wrong,
then we'll see...

* Part 1 solved!

* Part 2 complete!

* Clean up and finish up.

Co-authored-by: Kyle Sessions <ksessions@quantcast.com>